+++ src/data_structures/APR/access/GenAccess.hpp
@@ -13,18 +13,18 @@ public:
 
     GenInfo* genInfo;
 
-    uint64_t level_max() const { return genInfo->l_max; }
-    uint64_t level_min() const { return genInfo->l_min; }
+    int level_max() const { return genInfo->l_max; }
+    int level_min() const { return genInfo->l_min; }
 
-    uint64_t spatial_index_x_max(const unsigned int level) const { return genInfo->x_num[level]; }
-    uint64_t spatial_index_y_max(const unsigned int level) const { return genInfo->y_num[level]; }
-    uint64_t spatial_index_z_max(const unsigned int level) const { return genInfo->z_num[level]; }
+    int spatial_index_x_max(const unsigned int level) const { return genInfo->x_num[level]; }
+    int spatial_index_y_max(const unsigned int level) const { return genInfo->y_num[level]; }
+    int spatial_index_z_max(const unsigned int level) const { return genInfo->z_num[level]; }
 
-    uint64_t x_num(const unsigned int level) const { return genInfo->x_num[level]; }
-    uint64_t y_num(const unsigned int level) const { return genInfo->y_num[level]; }
-    uint64_t z_num(const unsigned int level) const { return genInfo->z_num[level]; }
+    int x_num(const unsigned int level) const { return genInfo->x_num[level]; }
+    int y_num(const unsigned int level) const { return genInfo->y_num[level]; }
+    int z_num(const unsigned int level) const { return genInfo->z_num[level]; }
 
-    unsigned int org_dims(int dim) const { return genInfo->org_dims[dim]; }
+    int org_dims(int dim) const { return genInfo->org_dims[dim]; }
 
 };
